AI-Service/requirements.txt
# AI Service Requirements
# Python dependencies for CV parsing and OpenAI integration

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI
openai==1.54.0
httpx==0.27.0

# LangFuse (AI Observability - Optional)
langfuse==2.27.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# DOCX Processing
python-docx==1.1.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
